% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discret.R
\name{discret}
\alias{discret}
\title{Discretization of a continuous variable}
\usage{
discret(x, classes = 5, type = 7, ...)
}
\arguments{
\item{x}{numerical vector}

\item{classes}{a number of a numeric vector of two or more unique cut points giving the number of intervals into which x will be cut}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used. For more information see the quantile fuction}

\item{...}{other argumnt to \code{\link{discret}} function}
}
\value{
a vector
}
\description{
Transform a continous variable into a discret variable.
}
\details{
for later..
}
\examples{
x <- rnorm(1000)
dx <- discret(x,4)

}
