% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergodic.R
\name{ergodic}
\alias{ergodic}
\title{Ergodic distribution of a Marov's probability transition matrix}
\usage{
ergodic(m)
}
\arguments{
\item{m}{Markov probability transition matrix}
}
\value{
a vector coantaning the ergodic distribution.
}
\description{
Calulates the ergodic distribution of a Marov's probability transition matrix
}
\details{
for later...
}
\examples{
t <- matrix(c(.5,.25,.25,.5,0,.5,.25,.25,.5),3,3,byrow=TRUE)
ergodic(t)

}
