% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guerry.R
\docType{data}
\name{Guerry}
\alias{Guerry}
\title{Data From A.-M. Guerry, "Essay On The Moral Statistics Of France"}
\format{A Spatial Object conatianing the original Guerry's dataset with two new variables \describe{
 \item{pc1}{First principal component of Crimes Against Persons, Crimes Against Property, Literacy, Donations, Infants Born out of Wedlock, and Suicides data}
 \item{pc2}{Second principal component of Crimes Against Persons, Crimes Against Property, Literacy, Donations, Infants Born out of Wedlock, and Suicides data}
}}
\usage{
data(Guerry)
}
\description{
Data extract from \code{\link[Guerry]{Guerry}} R package
}
\examples{
data(Guerry)
}
\keyword{dataset}
