% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo.test.R
\name{homo.test}
\alias{homo.test}
\title{Test of time homogeneity}
\usage{
homo.test(x, classes = 5, pr = 3)
}
\arguments{
\item{x}{numerical matrix of n spatial unit ans t time periods}

\item{classes}{a number of a numeric vector of two or more unique cut points giving the number of intervals into which x will be cut}

\item{pr}{number of subperiods in which to divide the entire sample}
}
\value{
A list coantaning the Q statistic and the LR statistic.
}
\description{
Computes the Test of time homogeneity  based on Bickenbach and Bode (2003)
}
\details{
for later.....
}
\examples{
data(us48)
data <- as.data.frame(us48)
pci <- data[,10:90]
rpci <- pci/matrix(1,dim(pci))\%*\%colMeans(pci)
homo.test(rpci,pr=5)

}
\references{
Bickenbach, F. and Bode, E. 2003. Evaluating the Markov Property in Studies of Economic Convergence, International Regional Science Review, vol. 26, no. 3, 363–92
}
