% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.d.R
\name{join.d}
\alias{join.d}
\title{Test the independence in the dynamics of a variable and its neighbors}
\usage{
join.d(x, W)
}
\arguments{
\item{x}{numerical matrix of n spatial unit ans t time periods}

\item{W}{an objet of listw class.}
}
\value{
a list coantaning \itemize{
   \item the Chi square statistcas and its p value
   \item The LISA markov transition matrix ander the null
   }
}
\description{
Performes a Chi square test for test that dynamics of a variable is independent of dynamics of its neighbors.
}
\details{
The test decompose the LISA Markov chain in a pair of chains, one for the city and other for the neighbors, each chain has two states H and L and under the null of independence test the co-movement of the chains.
}
\examples{
data(us48)
data <- as.data.frame(us48)
w1queen <- nb2listw(poly2nb(us48))
join.d(data[,10:90],w1queen)

}
