% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisamkv.R
\name{lisamkv}
\alias{lisamkv}
\title{Markov for Local Indicators of Spatial Association}
\usage{
lisamkv(x, W)
}
\arguments{
\item{x}{numerical matrix of n spatial unit ans t time periods}

\item{W}{an objet of listw class.}
}
\value{
a list cantaning three object
\describe{
  \item{move}{a data frame indicating which type of LISA transition occurred}
  \item{lisamatrix}{markov LISA transition matrix}
  \item{p.lisamatrix}{markov probability LISA transition matrix}
}
}
\description{
Compute the  Markov transition matrix for Local Indicators of Spatial Association
}
\details{
For later...
}
\examples{
data(us48ls)
ll <- lisamkv(data[,9:48],w1rook)

}
