% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\docType{data}
\name{mexico}
\alias{mexico}
\title{Mexican states regional income 1940-2000}
\format{A data frame contaning...}
\source{
\url{https://github.com/pysal/pysal/tree/master/pysal/examples/mexico}
}
\usage{
data(mexico)
}
\description{
Data from Mexican states regional income from 1940-2000 used in Rey and Guiterrez (2010)
}
\examples{
data(mexico)
str(mexico)
}
\references{
Rey, S.J. and M.L. Sastre Gutierrez. (2010) "Interregional inequality dynamics in Mexico." Spatial Economic Analysis, 5: 277-298.
}
\keyword{dataset}
