% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\title{Univariable and Bivariable Local Moran's I}
\usage{
moran(x, W, nsim = NULL, type = "uni", nbcom = length(W$neighbours),
  ...)
}
\arguments{
\item{x}{a vector, matrix or data frame containing the variables and spatial units}

\item{W}{a \code{listw} object}

\item{nsim}{number of random permutation used to the compute the pseudo p value. By default it is NULL}

\item{type}{a character indicating the local Moran's I to be compute. If it is set as "uni" the standard local Moran's I statistic is calulated. Set type as "multi" to cumpute the bivariable local Moran's I.}

\item{nbcom}{number of comparisons use in the Bonferroni multiple comparisons correction. By default it is set as the number of spatial unit to use.}

\item{...}{other argument to \code{quad} function. See \code{\link{quad}} for more infomation.}
}
\value{
a data frame or a list of data frames (more explanation in Details later)
}
\description{
Compute the Univariable (Anselin,1995) and Bivariable (cite??) Local Moran's I
}
\details{
later...
}
\examples{
data(Guerry)
w1queen<-nb2listw(poly2nb(Guerry))
pcc <- cbind(Guerry$pc1,Guerry$pc2)
\dontrun{ e4 <- moran(Guerry$pc1,w1queen,nsim=999)
e6 <- moran(pcc,w1queen,type="multi",nsim=999,geoda=FALSE)}

}
\references{
Anselin, L. (1995). Local Indicators of Spatial Association—LISA. Geographical Analysis, 27(2), 93–115. \url{https://doi.org/10.1111/j.1538-4632.1995.tb00338.x}
}
