% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prais.R
\name{prais}
\alias{prais}
\title{Prais conditional mobility measure.}
\usage{
prais(x)
}
\arguments{
\item{x}{Markov probability transition matrix.}
}
\value{
a vector
}
\description{
Compute the Prais conditional mobility measure.
}
\details{
Prais' conditional mobility measure for a class is defined as:
    \eqn{ pr_i = 1 -  p_{i,i}}
}
\examples{
data(us48)
data <- as.data.frame(us48)
pci <- data[,10:90]
rpci <- pci/matrix(1,dim(pci))\%*\%colMeans(pci)
m<-mkv(rpci)
prais(m[[1]])

}
