% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad.R
\name{quad}
\alias{quad}
\title{Quandrant Allocator}
\usage{
quad(x, W, y, geoda = TRUE)
}
\arguments{
\item{x}{a vector}

\item{W}{a \code{listw} object.}

\item{y}{a vector, use it in case that the lag must be compute in another variable.}

\item{geoda}{logial.  If True use GeoDa quadrant scheme: HH=1, LL=2, LH=3, HL=4. If False use PySAL quadrant Scheme: HH=1, LH=2, LL=3, HL=4}
}
\value{
a vector
}
\description{
Compute the position of a spatial unit and its neihbors in the MOran Scatter Plot
}
\details{
explain here what is the geoda
}
\examples{
data(Guerry)
quad(pc1,w1queen)

}
