% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorrock.R
\name{shorrock}
\alias{shorrock}
\title{Shorrock's mobility measure}
\usage{
shorrock(x)
}
\arguments{
\item{x}{Markov probability transition matrix.}
}
\value{
a vector
}
\description{
Compute the Shorrock's mobility measure
}
\examples{
data(usincome)
m <- mkv(rpci)
shorrock(m[[1]])

}
