% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.lisamkv.R
\name{sig.lisamkv}
\alias{sig.lisamkv}
\title{Markov for Local Indicators of Spatial Association including a significant state}
\usage{
sig.lisamkv(x, W, nsim = 999, ...)
}
\arguments{
\item{x}{numerical matrix of n spatial unit ans t time periods}

\item{W}{an objet of listw class.}

\item{nsim}{number of random spatial permutations for calculation of pseudo p-values, by default is set in 999}

\item{...}{other argument to \code{quad} function. See \code{\link{quad}} for more infomation.}
}
\value{
a list cantaning three object
\describe{
  \item{move}{a data frame indicating which type of LISA transition occurred}
  \item{lisamatrix}{markov LISA transition matrix}
  \item{p.lisamatrix}{markov probability LISA transition matrix}
}
}
\description{
Compute the  Markov transition matrix for Local Indicators of Spatial Association based in the significance of the local moran indicator
}
\details{
first the the Local Moran indicator is compute useing the \code{nsim} number of permutation. The non significant indication is considere as an state of the markow chain.
}
\examples{
\dontrun{
data(us48)
data <- as.data.frame(us48)
w1queen <- nb2listw(poly2nb(us48))
ll <- sig.lisamkv(data[,10:90],w1queen,999,geoda=FALSE)
}
}
