% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mkv.R
\name{sp.mkv}
\alias{sp.mkv}
\title{Spatial Markov transition probability matrix}
\usage{
sp.mkv(m, W, classes = 5, fixed = TRUE)
}
\arguments{
\item{m}{numerical matrix of n spatial unit ans t time periods}

\item{W}{an objet of listw class.}

\item{classes}{a number of a numeric vector of two or more unique cut points giving the number of intervals into which x will be cut}

\item{fixed}{logical, if it is TRUE the data are pooled over space and time and the quintiles calculated for the pooled data}
}
\value{
a list contaning the markov's trantiiona matrix and the markov's transition probability matrix
}
\description{
Compute the Spatial Markov transition probability matrix (Rey,2001)
}
\details{
for later..
}
\examples{
data(usincome)
sp.mkv(rpci,w)

}
\references{
Rey, S.J. (2001) “Spatial empirics for economic growth and convergence”, 34 Geographical Analysis, 33, 195-214.
}
