% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.tau.R
\name{sp.tau}
\alias{sp.tau}
\title{Global Indicators of Mobility Association (GIMA)}
\usage{
sp.tau(x, y, W, perm = NULL)
}
\arguments{
\item{x}{rank variable}

\item{y}{rank variable.}

\item{W}{an object of class \code{listw}}

\item{perm}{number of random spatial permutations for calculation of pseudo p-values, the default value is NULL.}
}
\value{
a vector coantaning \itemize{
   \item The GIMA statistic value
   \item The GIMA's pseudo p value, only available for perm >0
  \item The number of spatial Concordant pairs
  \item The number of spatial Discordant pairs
   \item The Tau statistic value
   \item The tau's p value
  \item The number of Concordant pairs
	  \item The number of Discordant pairs
  \item The number of extra x pairs. An extra x pair is a pair which \eqn{sgn(x_i - x_j)=0}
  \item The number of extra y pairs. An extra y pair is a pair which \eqn{sgn(y_i - y_j)=0}
   }
}
\description{
Compute the Global Indicators of Mobility Association index Rey (2016)
}
\details{
The Global Indicators of Mobility Association (GIMA) is based on the Spatial Tau indicator (Rey,2004). The implementation is a two step algorith based on the Rey's implementation (Rey,2014)
}
\examples{
data(mexico)
n <- nrow(mexico)
w <- matrix(0,n,n)
for(i in 1:n){
	w[,i] <- 1*(mexico$Regime[i]==mexico$Regime)
}
diag(w) <- 0
W <- mat2listw(w,style="B")
a <- lapply(1:(ncol(mexico)-2),function(i) sp.tau(mexico[,i],mexico[,(i+1)],W,999))
b<-do.call(rbind.data.frame,a)
colnames(b) <- names(a[[1]])
b <- round(b,3)


}
