% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.st.R
\name{st.st}
\alias{st.st}
\title{Steady State distribution of a Marov's probability transition matrix}
\usage{
st.st(m)
}
\arguments{
\item{m}{Markov probability transition matrix}
}
\value{
a vector coantaning the steady state distribution.
}
\description{
Calulates the steady stae distribution of a Marov's probability transition matrix
}
\details{
for later...
}
\examples{
t <- matrix(c(.5,.25,.25,.5,0,.5,.25,.25,.5),3,3,byrow=TRUE)
st.st(t)

}
