% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tau.R
\name{tau}
\alias{tau}
\title{Kendall rank correlation coefficient}
\usage{
tau(x, y)
}
\arguments{
\item{x}{rank variable}

\item{y}{rank variable.}
}
\value{
a vector coantaning \itemize{
   \item The Tau statistic value
   \item The tau's p value
  \item The number of concordant pairs
	  \item The number of Discordant pairs
  \item The number of extra x pairs. An extra x pair is a pair which \eqn{sgn(x_i - x_j)=0}
  \item The number of extra y pairs. An extra y pair is a pair which \eqn{sgn(y_i - y_j)=0}
   }
}
\description{
Compute the Kendall's tau rank correlation coefficient index
}
\details{
Kendall’s Tau is a non-parametric measure of relationships between columns of ranked data. The implemtation is based on Rey(2014),
}
\examples{
a<-c(12,2,1,12,2)
b<-c(1,4,7,1,0)
cor(a,b,method="kendall")
tau(a,b)


}
