% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u48.R
\docType{data}
\name{us48}
\alias{us48}
\title{Nominal per capita 48 US states time series 1929-2009}
\format{
an Spatial object
}
\source{
\url{https://github.com/pysal/pysal/tree/master/pysal/examples/us_income}
}
\usage{
data(us48)
}
\description{
Data used in spatial_dynamics module of PySAL library
}
\keyword{datasets}
