% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.LISA.R
\name{d.LISA}
\alias{d.LISA}
\title{Directional LISA}
\usage{
d.LISA(
  x0,
  x1,
  W,
  Regime = NULL,
  k = 8,
  mean.rel = FALSE,
  nsim = NULL,
  arrow = TRUE,
  only = NULL
)
}
\arguments{
\item{x0}{a vector containing the variable at initial period of analysis}

\item{x1}{a vector containing the variable at final period of analysis}

\item{W}{an object of class \code{listw}}

\item{Regime}{a numeric vector containing the regime to the spatial unit belongs}

\item{k}{a scalar in c(4,8) indicating number of circular sectors in rose diagram, by default it is set as 8}

\item{mean.rel}{a logical, Is the data mean relative?. By default it is FALSE}

\item{nsim}{number of random spatial permutations for calculation of pseudo p-values, the default value is NULL.}

\item{arrow}{Logical. Do you want to plot the arrows in the standardised plot?. If it is set as FALSE the arrow head is plot as a point.}

\item{only}{numerical indicating the spatial unit to be considered in the plot, the NULL value imply the use of all the spatial unit in the plot.}
}
\value{
a list contaning \describe{
    \item{"Lisa"}{A vector scatterplot showing the Directinal co-movement of a spatial unit and its neighbors}
    \item{"rose"}{A rose diagram of the LISA}
    \item{"p.rose"}{pseudo p-values of the direction in the rose diagram. Only available when nsimm is not NULL}
    \item{"data"}{a data frame contening the data used for the lisa scatterplot}
    \item{"counts"}{a data frame containing the information used for the rose diagram}
    \item{"p.value"}{a data frame containing the information used in the pseudo p-value graph. Only available when nsim is not NULL}
 }
}
\description{
Compute the origin standardized movement of a spatial unit and its neighbors and the pseudo p-value of the directional co-movement.
}
\details{
For later...
}
\examples{
data(us48)
w1queen <- nb2listw(poly2nb(us48))
t0<-us48$X1969/mean(us48$X1969)
t1 <-us48$X2009/mean(us48$X2009)
Regime <-us48$SUB_REGION
ok <- d.LISA(t0,t1,w1queen,Regime,k=8,nsim=999)

}
\references{
Rey, S. J., Murray, A. T., & Anselin, L. (2011). Visualizing regional income distribution dynamics. Letters in Spatial and Resource Sciences, 4(1), 81–90.
}
