% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo.test.R
\name{homo.test}
\alias{homo.test}
\title{Test for homogeneity of Markov transition probabilities across regimes.}
\usage{
homo.test(x, W, classes = 5, fixed = TRUE)
}
\arguments{
\item{x}{numerical matrix of n spatial unit ans t time periods}

\item{W}{an objet of listw class.}

\item{classes}{a number of a numeric vector of two or more unique cut points giving the number of intervals into which x will be cut}

\item{fixed}{logical, if it is TRUE the data are pooled over space and time and the quintiles calculated for the pooled data}
}
\value{
A list coantaning the Q statistic, the LR statistic and matrix use as null hypotesis in the test.
}
\description{
Performs the homogenity across space test for spatial markov trasntion matrix basis on Rey et al, (2016)
}
\details{
For later...
}
\examples{
data(us48)
data <- as.data.frame(us48)
pci <- data[,10:90]
rpci <- pci/matrix(1,dim(pci))\%*\%colMeans(pci)
w1queen <- nb2listw(poly2nb(us48))
homo.test(rpci,w1queen)

}
\references{
S. J. Rey, W. Kang, and L. Wolf (2016) “The properties of tests for spatial effects in discrete Markov chain models of regional income distribution dynamics,” Journal of  Geographical Systems, vol. 18, no. 4, pp. 377–398.
}
