% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisamkv.R
\name{lisamkv}
\alias{lisamkv}
\title{Markov for Local Indicators of Spatial Association}
\usage{
lisamkv(x, W, ...)
}
\arguments{
\item{x}{numerical matrix of n spatial units and t time periods}

\item{W}{an objet of listw class.}

\item{...}{other argument to \code{quad} function. See \code{\link{quad}} for more infomation.}
}
\value{
a list cantaining three objects
\describe{
  \item{move}{a data frame indicating which type of LISA transition occurred}
  \item{lisamatrix}{markov LISA transition matrix}
  \item{p.lisamatrix}{markov probability LISA transition matrix}
}
}
\description{
Compute the  Markov transition matrix for Local Indicators of Spatial Association
}
\details{
For later...
}
\examples{
data(us48)
data <- as.data.frame(us48)
w1queen <- nb2listw(poly2nb(us48))
ll <- lisamkv(data[,10:90],w1queen)

}
