% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorrock.R
\name{shorrock}
\alias{shorrock}
\title{Shorrock's mobility measure}
\usage{
shorrock(x)
}
\arguments{
\item{x}{Markov probability transition matrix.}
}
\value{
a vector
}
\description{
Compute the Shorrock's mobility measure
}
\examples{
data(us48)
data <- as.data.frame(us48)
pci <- data[,10:90]
rpci <- pci/matrix(1,dim(pci))\%*\%colMeans(pci)
m <- mkv(rpci)
shorrock(m[[1]])

}
